pyproject.toml
[virtualenvs]
in-project = true

[tool.poetry]
name = "zoom-lamp"
version = "0.1.0"
description = ""
authors = ["Andrew Cleasby <andrew.cleasby@nydig.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
paho-mqtt = "^1.6.1"
watchdog = "^2.1.6"
#PySide6 = "^6.2.4"
#shiboken6 = "^6.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
